+++ server/faker-sym.h
@@ -567,6 +567,16 @@ FUNCDEF1(const GLubyte *, glGetString, GLenum, name, glGetString)
 FUNCDEF2(const GLubyte *, glGetStringi, GLenum, name, GLuint, index,
 	glGetStringi)
 
+#ifdef GL_VERSION_4_5
+
+VFUNCDEF2(glNamedFramebufferDrawBuffer, GLuint, framebuffer, GLenum, buf,
+	glNamedFramebufferDrawBuffer)
+
+VFUNCDEF3(glNamedFramebufferDrawBuffers, GLuint, framebuffer, GLsizei, n,
+	const GLenum *, bufs, glNamedFramebufferDrawBuffers)
+
+#endif
+
 VFUNCDEF0(glPopAttrib, glPopAttrib)
 
 VFUNCDEF4(glViewport, GLint, x, GLint, y, GLsizei, width, GLsizei, height,